Update Python to 2.7.15 and update nearly all pips (#132)
* requirements.txt for python 2.7.15 and current pips

* Update mostly to the latest pip packages.
* Use pip==9.0.3 since pip==10.0.0 breaks cobra test.
* Use FireWorks==1.6.0 since 1.7.2 clashes with cobra over the needed
version of ruamel.yaml.
* Add pytest and pytest-benchmark for cobra test and because pytest
is super useful. pytest-benchmark looks useful, too.

* --enable-shared when installing python; back off to cvxopt 1.1.7

* Installing Theano on Sherlock requires a shared library python
(libpython.so).
* cvxopt 1.2.0 installs on OSX but not on Sherlock, so back off to
1.1.7 for now.

* convert floats to int for recent numpy; back off some pips

* For compatibility with recent versions of numpy,
bulk_objects_container.py needs to convert some float values (like
array means) to integers in countsInc() and countsDec(). Add failing
tests, then the fixes. The fixes support scalar, list, and ndarray
values.

* Revert to bokeh 0.11.1 until we adjust to API changes.
[TODO] Adjust and update.

* Back off cvxpy to 0.4.8 since the 0.4.11 change summary looks like
it’s a work in progress.

* The previous commit backed off cvxopt to 1.1.7 because 1.2.0
wouldn’t install on Sherlock. [TODO] Debug that.

* support any bulk object counts dtype

* do dtype-casting in BulkMolecules like in BulkObjects to support the newer version of NumPy.

I added a failing test case before making the fix to ensure that the fix
works, but its code to construct test data could be based on more
understanding than this hack based on guesses and experimentation.

* fix `chromosomeIndex[a_float_value]`

* Change distance metric from correlation to default (euclidean) in metabolismFullFluxes analysis

* use np.asarray() to convert count values instead of `np.array()` to save making a copy when `values` already has the right dtype.

numpy.testing.assert_equal() actually prints good failure info as long as one or both of the values are numpy arrays. It also simplifies the test code.

* revert matplotlib to 1.4.3 for now; pip bug fix updates

* Update 2 analysis scripts for current pips

proteinMassFractionFoldChange: Fix "ValueError: range parameter must be finite."

It still prints some warnings: "RuntimeWarning: Mean of empty slice."

ribosomeProduction: Fix "TypeError: `bins` must be an integer, a string, or an array".

We DEFINITELY need to make these command lines usable, e.g. set just the changing part of the base path, default everything else, and rationalize "simOutDir" vs. "seeOutDir".

* fix growthAffectingPolymerases; newer matplotlib

* growthAffectingPolymerases: Adapt "TypeError: `bins` must be an integer, a string, or an array" for recent pips.

* Update to matplotlib==1.5.3. This fixes "FutureWarning: elementwise comparison failed; returning scalar instead, but in the future will perform elementwise comparison".

* fix proteinMassFractionFoldChange for newer numpy using `massFractionRatio > 0` to filter out negative and NaN values. `np.logical_or()` only takes two input values. `|` and `&` are more convenient for boolean arrays.

Separate the variable `massFractionRatio_logs` from `massFractionRatio_realValues`.

Comment out unused code that generates  needless "Mean of empty slice" warnings.

* fix errors and warnings from matplotlib

* More histogram "`bins` must be an integer, a string, or an array'" errors.
* More set_color_cycle(colors) deprecation warnings.

* revert to set_color_cycle() until everyone gets up on the new matplotlib

Temporarily go back to the future-warning so the code doesn't break with the old matplotlib. Plan to `git revert` this change after people can adopt the new matplotlib.

* remove debug print stmt; allow newer virtualenv

+++ models/ecoli/analysis/multigen/growthAffectingPolymerases.py
@@ -12,6 +12,7 @@ import matplotlib.gridspec as gridspec
 from models.ecoli.analysis.AnalysisPaths import AnalysisPaths
 from wholecell.io.tablereader import TableReader
 import wholecell.utils.constants
+from wholecell.analysis.analysis_tools import exportFigure
 from wholecell.utils import units
 import cPickle
 
@@ -217,7 +218,7 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 		ax1.axvline(x = time.asNumber(units.min).max(), linewidth=2, color='k', linestyle='--')
 
 		hist_doublingTime = removeNanReshape(doublingTime.asNumber(units.min))
-		nbins = np.ceil(np.sqrt(hist_doublingTime.size))
+		nbins = int(np.ceil(np.sqrt(hist_doublingTime.size)))
 		ax1_1.hist(hist_doublingTime, nbins, (hist_doublingTime.mean() - hist_doublingTime.std() / 2, hist_doublingTime.mean() + hist_doublingTime.std() / 2))
 
 		# Plot RNAP active fraction
@@ -227,7 +228,7 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 		ax2.set_ylabel("Fraction active\nRNAP")
 
 		hist_fractionRnapActive = removeNanReshape(fractionRnapActive)
-		nbins = np.ceil(np.sqrt(hist_fractionRnapActive.size))
+		nbins = int(np.ceil(np.sqrt(hist_fractionRnapActive.size)))
 		ax2_1.hist(hist_fractionRnapActive, nbins, (hist_fractionRnapActive.mean() - hist_fractionRnapActive.std(), hist_fractionRnapActive.mean() + hist_fractionRnapActive.std()))
 
 		# Plot RNAP active and total counts
@@ -357,7 +358,6 @@ def main(seedOutDir, plotOutDir, plotOutFileName, simDataFile, validationDataFil
 
 	fig.subplots_adjust(hspace=.5, wspace = 0.3)
 
-	from wholecell.analysis.analysis_tools import exportFigure
 	exportFigure(plt, plotOutDir, plotOutFileName,metadata)
 	plt.close("all")
